Fixed cohesion error in update system
This fix implements a change the code which displays to the user which version they're utilizing, such as a release or development version. This error was reading the update channel byte in the SFDATA file which can return incorrect results if the user is utilizing one version that has their update channel switch.